upload file function

+++ app/Http/Controllers/PostsController.php
@@ -3,12 +3,23 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
 // use Illuminate\View\Middleware\ShareErrorsFromSession;
 
 class PostsController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -44,14 +55,34 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // Handle file upload
+        if ($request->hasFile('cover_image')) {
+            // Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get only filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get only file extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // File name to store
+            $filenameToStore = $filename.'_'.time().'.'.$extension;
+            // Upload the image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $filenameToStore);
+        }
+        else {
+            $filenameToStore = 'noimage.jpg';
+        }
 
         // return 123;
         // Create post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $filenameToStore;
         $post->save();
 
         // Redirect
@@ -80,6 +111,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // check user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -97,11 +134,28 @@ class PostsController extends Controller
             'body' => 'required'
         ]);
 
+        // Handle file upload
+        if ($request->hasFile('cover_image')) {
+            // Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get only filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get only file extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // File name to store
+            $filenameToStore = $filename.'_'.time().'.'.$extension;
+            // Upload the image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $filenameToStore);
+        }
+    
         // return 123;
         // Create post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if ($request->hasFile('cover_image')){
+            $post->cover_image = $filenameToStore;
+        }
         $post->save();
 
         // Redirect
@@ -117,6 +171,17 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // check user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
+        if($post->cover_image != 'noimage.jpg'){
+            // Delete image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
